use configuration matrix
