+++ mod-tool/src/test/scala/org/apalachemc/integration/README.md
@@ -28,6 +28,40 @@ Use `workspace.write` for test-local input and `workspace.filename("Spec.tla")` 
 `expectOutput`. Assertions inspect stdout by default; pass `isStderr = true` to inspect stderr.
 `expectOutput` supports a whole-line `...` wildcard matching any number of lines.
 
+## Configuration matrix
+
+Suites use the configuration-independent `GENERAL` worker by default. This is appropriate for commands such as
+`parse`, `typecheck`, and `version` that do not use an SMT solver. Checker suites declare the exact configurations
+under which all of their tests are valid:
+
+```scala
+import org.apalachemc.integration.framework.IntegrationTestConfiguration.{ARRAYS_Z3, OOPSLA19_CVC5, OOPSLA19_Z3}
+import org.apalachemc.integration.framework.IntegrationTestConfiguration
+
+class CheckCommandTest extends IntegrationTestBase {
+  override protected val supportedConfigurations: Set[IntegrationTestConfiguration] =
+    Set(OOPSLA19_Z3, OOPSLA19_CVC5, ARRAYS_Z3)
+}
+```
+
+The available configuration IDs are:
+
+- `general`
+- `oopsla19-z3`
+- `oopsla19-cvc5`
+- `arrays-z3`
+
+`arrays-cvc5` is not supported. Put scenarios that work with every checker configuration in a shared suite. Put
+OOPSLA19-only scenarios in a suite supporting `OOPSLA19_Z3` and `OOPSLA19_CVC5`, and solver-specific scenarios in a
+suite supporting one configuration. Compatibility is a suite property; tags remain available for independent
+categories such as temporal tests.
+
+Each selected configuration runs in its own forked JVM. Configuration workers run in parallel, while the tests in
+one worker run sequentially. The checker workers set `SMT_SOLVER` and `SMT_ENCODING` for both in-process commands
+and commands that use a fresh JVM.
+
+## Execution mode
+
 Tests run Tool in-process by default. Force a particular test to use a fresh JVM with `Forked`:
 
 ```scala
@@ -41,10 +75,14 @@ test("exercise process isolation", Forked) {
 Run the suites with:
 
 ```sh
-sbt tool/cliIntegrationTest
+make scala-integration
+APALACHE_CLI_TEST_CONFIGS=oopsla19-cvc5 sbt tool/cliIntegrationTest
+APALACHE_CLI_TEST_CONFIGS=oopsla19-z3,arrays-z3 sbt tool/cliIntegrationTest
 APALACHE_CLI_TEST_MODE=forked sbt tool/cliIntegrationTest
 sbt 'tool/CliIntegration/testOnly org.apalachemc.integration.ParseCommandTest -- -z "empty module"'
 ```
 
-Set `APALACHE_CLI_TEST_TIMING=true` to print the elapsed time of each Tool invocation. Suite-level
-parallel execution is currently disabled because in-process runs temporarily replace JVM-global state.
+`APALACHE_CLI_TEST_CONFIGS` filters the worker IDs; if omitted, all four workers run. A `testOnly` selection is still
+expanded across the selected workers, and the suite itself skips workers it does not support.
+
+Set `APALACHE_CLI_TEST_TIMING=true` to print the active configuration and elapsed time of each Tool invocation.
