+++ mod-tool/src/test/scala/org/apalachemc/integration/framework/IntegrationTestBase.scala
@@ -1,14 +1,25 @@
 package org.apalachemc.integration.framework
 
 import org.scalatest.BeforeAndAfterEach
-import org.scalatest.Outcome
+import org.scalatest.{Args, Outcome, Status, SucceededStatus}
 import org.scalatest.funsuite.AnyFunSuite
 
 /** Supplies each command suite with an isolated workspace and selected Tool runner. */
 abstract class IntegrationTestBase extends AnyFunSuite with BeforeAndAfterEach {
+  import IntegrationTestConfiguration.GENERAL
+
   private val reportTimings = sys.env.get("APALACHE_CLI_TEST_TIMING").exists(_.equalsIgnoreCase("true"))
   private var currentRunner: ToolRunner = _
   private var currentWorkspace: TestWorkspace = _
+
+  /** Configurations under which this entire suite is valid. */
+  protected def supportedConfigurations: Set[IntegrationTestConfiguration] = Set(GENERAL)
+
+  /** Configuration selected for this worker process. */
+  final protected def activeConfiguration: IntegrationTestConfiguration = IntegrationTestConfiguration.active
+
+  /** Distinguishes reports emitted by parallel configuration workers. */
+  final override def suiteName: String = s"${super.suiteName} [${activeConfiguration.id}]"
 
   /** Returns the isolated workspace belonging to the current test. */
   final protected def workspace: TestWorkspace = {
@@ -21,7 +32,9 @@ abstract class IntegrationTestBase extends AnyFunSuite with BeforeAndAfterEach {
     require(currentRunner != null, "The Tool runner is only available while a test is running")
     val result = currentRunner.run(workspace, arguments)
     if (reportTimings) {
-      println(f"CLI_TEST_COMMAND_TIME_MS ${result.elapsed.toNanos / 1000000.0}%.3f")
+      println(
+          f"CLI_TEST_COMMAND_TIME_MS configuration=${activeConfiguration.id} " +
+            f"elapsed=${result.elapsed.toNanos / 1000000.0}%.3f")
     }
     result
   }
@@ -29,6 +42,17 @@ abstract class IntegrationTestBase extends AnyFunSuite with BeforeAndAfterEach {
   /** Runs a command with its output directory set to the current workspace. */
   final protected def runWithOutDir(command: String, arguments: String*): CommandResult =
     run((Seq(command, s"--out-dir=${workspace.outDir}") ++ arguments): _*)
+
+  /** Silently omits this suite from workers whose configuration it does not support. */
+  override protected def runTests(testName: Option[String], args: Args): Status = {
+    IntegrationTestConfiguration.validateSupported(supportedConfigurations)
+    if (supportedConfigurations.contains(activeConfiguration)) {
+      activeConfiguration.validateEnvironment(sys.env)
+      super.runTests(testName, args)
+    } else {
+      SucceededStatus
+    }
+  }
 
   /** Selects the runner indicated by the current test's tags. */
   override protected def withFixture(test: NoArgTest): Outcome = {
